Retire dependency-evidence command
Remove the standalone command and positional library fallback while preserving normalized evidence through depends and reserving the retired token with replacement guidance.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
