+++ docs/README.md
@@ -90,7 +90,7 @@ Package: System.Text.Json | Version: 10.0.0 | TFM: net10.0 | Library: lib/net10.
 | [Package Input and Dependency Evidence](design/package-dependency-evidence.md) | Common declarations and produced relationships across project, package, assets, and runtime inputs, preserving authorship, requested/resolved evidence, processing observations, equivalence, completion, and `InertString` containment. |
 | [Package Dependency Candidate Resolution](design/package-dependency-candidate-resolution.md) | Host-neutral composition from one normalized declaration to an exact source-authorized package candidate, typed failure, or incomplete source evidence. |
 | [Package Dependency Traversal](design/package-dependency-traversal.md) | Source-authorized exact package-manifest traversal with typed graph identity, root-relative depth, failures, completion, and shared CLI/Browser consumption. |
-| [Dependency Evidence CLI](design/dependency-evidence-cli.md) | CLI command, input binding, sections, count exactness, Markout and JSON lowering, diagnostics, and routing for normalized dependency evidence. |
+| [Dependency Evidence CLI](design/dependency-evidence-cli.md) | Historical contract for the retired `dependency-evidence` command; current behavior is owned by Dependency Inspection Command. |
 | [Authored Project Dependency Facts](design/authored-project-dependency-facts.md) | Host-neutral exact-project-XML target observations, package declarations, condition association, identity, incomplete evidence, failure, and containment. |
 | [Restored Project Dependency Facts](design/restored-project-dependency-facts.md) | Host-neutral `project.assets.json` declarations, resolved package graph, typed identity, completion, failure, and containment. |
 | [Runtime Dependency Facts](design/runtime-dependency-facts.md) | Host-neutral `.deps.json` runtime-target package nodes, package relationships, identity, completion, failure, and containment. |
