+++ docs/design/dependency-evidence-cli.md
@@ -3,17 +3,20 @@
 This document owns the command, input-binding, section, projection, and
 presentation contract for the CLI consumer tracked by #5534.
 
-**Status:** implementation target.
+**Status:** historical implementation contract. The command was retired by
+[#5995](https://github.com/richlander/dotnet-inspect/issues/5995);
+[Dependency Inspection Command](dependency-inspection-command.md) is the sole
+current CLI owner.
 
 > [!NOTE]
 > [Dependency Inspection Command](dependency-inspection-command.md) defines the
-> target convergence of this command and `depends` into one asset-driven
-> `depends` operation. This document remains the current implementation
-> contract until that migration and the explicit retirement of
-> `dependency-evidence` land. Its prior conclusion that heterogeneous root
+> convergence of this command and `depends` into one asset-driven
+> `depends` operation. This document records the retired implementation
+> contract. Its prior conclusion that heterogeneous root
 > cardinality requires a separate operation is superseded for that target:
 > the unified command closes one or several roots into one dependency document
-> and one per-root completion model.
+> and one per-root completion model. Do not use this document as current CLI
+> guidance.
 
 ## Owner and claim
