+++ src/DotnetInspect.Cli/CommandLine/Commands/RouterCommandDefinition.cs
@@ -481,7 +481,6 @@ public static class RouterCommandDefinition
         "extensions",
         "implements",
         "depends",
-        DependencyEvidenceCommand.Name,
         "cache",
         "skill",
         DemoCommand.Name
