+++ src/DotnetInspect.Cli/Program.cs
@@ -171,6 +171,12 @@ try
 
     var rootCommand = CommandLineBuilder.CreateRootCommand();
 
+    if (CommandLineBuilder.TryGetRemovedCommandError(args, out var removedCommandError))
+    {
+        CommandError.Write(removedCommandError!);
+        return 1;
+    }
+
     if (CommandLineBuilder.TryGetStaleArgumentError(
             args,
             rootCommand,
